Projects/AutoDeployment/requirements.txt
loguru==0.5.3
PyYAML==6.0
tqdm==4.64.1
